+++ pkg/controllers/build/function_controller.go
@@ -35,7 +35,6 @@ import (
 
 	buildv1alpha1 "github.com/projectriff/system/pkg/apis/build/v1alpha1"
 	kpackbuildv1alpha1 "github.com/projectriff/system/pkg/apis/thirdparty/kpack/build/v1alpha1"
-	"github.com/projectriff/system/pkg/printers"
 )
 
 // FunctionReconciler reconciles a Function object
@@ -75,7 +74,7 @@ func (r *FunctionReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	// check if status has changed before updating, unless requeued
 	if !result.Requeue && !equality.Semantic.DeepEqual(function.Status, originalFunction.Status) {
 		// update status
-		log.Info(fmt.Sprintf("Updating function status diff (-current, +desired): %s", cmp.Diff(originalFunction.Status, function.Status)))
+		log.Info("updating function status", "diff", cmp.Diff(originalFunction.Status, function.Status))
 		if updateErr := r.Status().Update(ctx, &function); updateErr != nil {
 			log.Error(updateErr, "unable to update Function status", "function", function)
 			return ctrl.Result{Requeue: true}, updateErr
@@ -104,13 +103,15 @@ func (r *FunctionReconciler) reconcile(ctx context.Context, log logr.Logger, fun
 	function.Status.MarkImageResolved()
 	function.Status.TargetImage = targetImage
 
-	// reconcile child image
-	childImage, err := r.reconcileChildImage(ctx, log, function)
+	// reconcile child kpack image
+	childImage, err := r.reconcileChildKpackImage(ctx, log, function)
 	if err != nil {
 		log.Error(err, "unable to reconcile child Image", "function", function)
 		return ctrl.Result{}, err
 	}
 	if childImage == nil {
+		// TODO resolve to a digest
+		function.Status.LatestImage = function.Status.TargetImage
 		function.Status.MarkBuildNotUsed()
 	} else {
 		function.Status.KpackImageName = childImage.Name
@@ -147,7 +148,7 @@ func (r *FunctionReconciler) resolveTargetImage(ctx context.Context, log logr.Lo
 	return image, nil
 }
 
-func (r *FunctionReconciler) reconcileChildImage(ctx context.Context, log logr.Logger, function *buildv1alpha1.Function) (*kpackbuildv1alpha1.Image, error) {
+func (r *FunctionReconciler) reconcileChildKpackImage(ctx context.Context, log logr.Logger, function *buildv1alpha1.Function) (*kpackbuildv1alpha1.Image, error) {
 	var actualImage kpackbuildv1alpha1.Image
 	if function.Status.KpackImageName != "" {
 		if err := r.Get(ctx, types.NamespacedName{Namespace: function.Namespace, Name: function.Status.KpackImageName}, &actualImage); err != nil {
@@ -171,18 +172,20 @@ func (r *FunctionReconciler) reconcileChildImage(ctx context.Context, log logr.L
 		return nil, err
 	}
 
-	// delete image if no longer needed
 	if desiredImage == nil {
-		if err := r.Delete(ctx, &actualImage); err != nil {
-			log.Error(err, "unable to delete kpack Image for Function", "image", actualImage)
-			return nil, err
+		// delete image if no longer needed
+		if actualImage.Name != "" {
+			if err := r.Delete(ctx, &actualImage); err != nil {
+				log.Error(err, "unable to delete kpack Image for Function", "image", actualImage)
+				return nil, err
+			}
 		}
 		return nil, nil
 	}
 
 	// create image if it doesn't exist
 	if function.Status.KpackImageName == "" {
-		log.Info(fmt.Sprintf("Creating kpack image spec: %s", printers.Pretty(desiredImage.Spec)))
+		log.Info("creating kpack image", "spec", desiredImage.Spec)
 		if err := r.Create(ctx, desiredImage); err != nil {
 			log.Error(err, "unable to create kpack Image for Function", "image", desiredImage)
 			return nil, err
@@ -199,7 +202,7 @@ func (r *FunctionReconciler) reconcileChildImage(ctx context.Context, log logr.L
 	image := actualImage.DeepCopy()
 	image.ObjectMeta.Labels = desiredImage.ObjectMeta.Labels
 	image.Spec = desiredImage.Spec
-	log.Info(fmt.Sprintf("Reconciling kpack image spec diff (-current, +desired): %s", cmp.Diff(actualImage.Spec, image.Spec)))
+	log.Info("reconciling kpack image", "diff", cmp.Diff(actualImage.Spec, image.Spec))
 	if err := r.Update(ctx, image); err != nil {
 		log.Error(err, "unable to update kpack Image for Function", "image", image)
 		return nil, err
